+++ operator_service/data_store.py
@@ -1,0 +1,153 @@
+import os
+import logging
+
+import psycopg2
+import simplejson as json
+
+
+def get_sql_status(agreement_id, job_id, owner):
+    # enforce strings
+    logging.debug("Start get_sql_status\n")
+    logging.debug("Connected\n")
+    params = dict()
+    select_query = '''
+    SELECT agreementId, workflowId, owner, status, statusText, 
+        extract(epoch from dateCreated) as dateCreated, 
+        extract(epoch from dateFinished) as dateFinished,
+        configlogURL, publishlogURL, algologURL, outputsURL, ddo, stopreq, removed 
+    FROM jobs WHERE 1=1
+    '''
+
+    if agreement_id is not None:
+        select_query = select_query + " AND agreementId=%(agreementId)s"
+        params['agreementId'] = str(agreement_id)
+
+    if job_id is not None:
+        select_query = select_query + " AND workflowId=%(jobId)s"
+        params['jobId'] = str(job_id)
+
+    if owner is not None:
+        select_query = select_query + " AND owner=%(owner)s"
+        params['owner'] = str(owner)
+
+    logging.debug(f'Got select_query: {select_query}')
+    logging.debug(f'Got params: {params}')
+
+    result = []
+    rows = _execute_query(select_query, params, 'get_sql_status', get_rows=True)
+    if not rows:
+        return
+
+    for row in rows:
+        temprow = dict()
+        temprow['agreementId'] = row[0]
+        temprow['jobId'] = row[1]
+        temprow['owner'] = row[2]
+        temprow['status'] = row[3]
+        temprow['statusText'] = row[4]
+        temprow['dateCreated'] = row[5]
+        temprow['dateFinished'] = row[6]
+        # temprow['configlogUrl']=row[7]
+        # temprow['publishlogUrl']=row[8]
+        temprow['algologUrl'] = row[9]
+        temprow['outputsUrl'] = row[10]
+        temprow['ddo'] = ''
+        temprow['did'] = ''
+        if row[11] is not None:
+            temprow['ddo'] = str(row[11])
+            ddo_json = json.loads(temprow['ddo'])
+            if 'id' in ddo_json:
+                temprow['did'] = ddo_json['id']
+        temprow['stopreq'] = row[12]
+        temprow['removed'] = row[13]
+        result.append(temprow)
+    return result
+
+
+def get_sql_jobs(agreement_id, job_id, owner):
+    logging.debug("Start get_sql_status\n")
+    logging.debug("Connected\n")
+    params = dict()
+    select_query = "SELECT workflowId FROM jobs WHERE 1=1"
+    if agreement_id is not None:
+        select_query = select_query + " AND agreementId=%(agreementId)s"
+        params['agreementId'] = str(agreement_id)
+    if job_id is not None:
+        select_query = select_query + " AND workflowId=%(jobId)s"
+        params['jobId'] = str(job_id)
+    if owner is not None:
+        select_query = select_query + " AND owner=%(owner)s"
+        params['owner'] = str(owner)
+    logging.error(f'Got select_query: {select_query}')
+    logging.error(f'Got params: {params}')
+    rows = _execute_query(select_query, params, 'get_sql_jobs', get_rows=True)
+    if not rows:
+        return
+
+    return [row[0] for row in rows]
+
+
+def create_sql_job(agreement_id, job_id, owner):
+    postgres_insert_query = """
+        INSERT 
+            INTO jobs 
+                (agreementId,workflowId,owner,status,statusText) 
+            VALUES 
+                (%s, %s, %s, %s, %s)"""
+    record_to_insert = (str(agreement_id), str(job_id), str(owner), 10, "Job started")
+    return _execute_query(postgres_insert_query, record_to_insert, 'create_sql_job')
+
+
+def stop_sql_job(execution_id):
+    postgres_insert_query = """UPDATE jobs SET stopreq=1 WHERE workflowId=%s"""
+    record_to_insert = (str(execution_id),)
+    return _execute_query(postgres_insert_query, record_to_insert, 'stop_sql_job')
+
+
+def remove_sql_job(execution_id):
+    postgres_insert_query = """ UPDATE jobs SET removed=1 WHERE workflowId=%s"""
+    record_to_insert = (str(execution_id),)
+    return _execute_query(postgres_insert_query, record_to_insert, 'remove_sql_job')
+
+
+def get_pg_connection_and_cursor():
+    try:
+        connection = psycopg2.connect(
+            user=os.getenv("POSTGRES_USER"),
+            password=os.getenv("POSTGRES_PASSWORD"),
+            host=os.getenv("POSTGRES_HOST"),
+            port=os.getenv("POSTGRES_PORT"),
+            database=os.getenv("POSTGRES_DB")
+        )
+        connection.set_client_encoding('LATIN9')
+        cursor = connection.cursor()
+        return connection, cursor
+    except (Exception, psycopg2.Error) as error:
+        logging.error(f'New PG connect error: {error}')
+        return None, None
+
+
+def _execute_query(query, record, msg, get_rows=False):
+    connection, cursor = get_pg_connection_and_cursor()
+    if not connection or not cursor:
+        return
+
+    try:
+        cursor.execute(query, record)
+        if get_rows:
+            result = []
+            row = cursor.fetchone()
+            while row is not None:
+                result.append(row[0])
+                row = cursor.fetchone()
+            return result
+        else:
+            connection.commit()
+
+    except (Exception, psycopg2.Error) as error:
+        logging.error(f'Got PG error in {msg}: {error}')
+    finally:
+        # closing database connection.
+        if connection:
+            cursor.close()
+            connection.close()
