+++ operator_service/routes.py
@@ -1,41 +1,32 @@
-import logging
-import uuid
-from configparser import ConfigParser
+import os
 from os import path
+import logging
 
 import kubernetes
-import os
-import psycopg2
-from flask import Blueprint, jsonify, request, Response
-from kubernetes import client, config
+from flask import Blueprint, jsonify, request
 from kubernetes.client.rest import ApiException
-import simplejson as json
+from ocean_utils.utils.utilities import generate_new_id
+
+from operator_service.config import Config
+from operator_service.data_store import create_sql_job, get_sql_status, get_sql_jobs, stop_sql_job, remove_sql_job
+from operator_service.kubernetes_api import KubeAPI
+from operator_service.utils import create_compute_job, check_required_attributes
 
 services = Blueprint('services', __name__)
 
 # Configuration to connect to k8s.
 if not path.exists('/.dockerenv'):
-    config.load_kube_config()
+    kubernetes.config.load_kube_config()
 else:
-    config.load_incluster_config()
+    kubernetes.config.load_incluster_config()
 
-# create instances of the API classes
-api_customobject = client.CustomObjectsApi()
-api_core = client.CoreV1Api()
-
-config_parser = ConfigParser()
-configuration = config_parser.read('/operator-service/config.ini')
-group = config_parser.get('resources', 'group')  # str | The custom resource's group name
-version = config_parser.get('resources', 'version')  # str | The custom resource's version
-namespace = config_parser.get('resources', 'namespace')  # str | The custom resource's namespace
-plural = config_parser.get('resources',
-                           'plural')  # str | The custom resource's plural name. For TPRs this would be
+config = Config()
 
 
 @services.route('/compute', methods=['POST'])
-def init_execution():
+def start_compute_job():
     """
-    Initialize the execution when someone call to the execute endpoint in brizo.
+    Create and start the compute job
     ---
     tags:
       - operation
@@ -112,62 +103,63 @@ def init_execution():
       400:
         description: Some error
     """
-    if 'agreementId' not in request.json:
-        logging.error(f'Missing agreementId')
-        return 'Missing agreementId', 400
-    if 'owner' not in request.json:
-        logging.error(f'Missing ownerId')
-        return 'Missing ownerId', 400
-    if 'stages' not in request.json:
+
+    data = request.json
+    required_attributes = [
+        'stages',
+        'agreementId',
+        'owner',
+    ]
+    msg, status = check_required_attributes(required_attributes, data, 'POST:/compute')
+    if msg:
+        return msg, status
+
+    agreement_id = data.get('agreementId')
+    owner = data.get('owner')
+    stages = data.get('stages')
+
+    if not stages:
         logging.error(f'Missing stages')
         return 'Missing stages', 400
-    if not request.json['stages']:
-        logging.error(f'Missing stages')
-        return 'Missing stages', 400
-    if 'algorithm' not in request.json['stages'][0]:
-        logging.error(f'Missing algorithm in stage')
-        return 'Missing algorithm in stages', 400
-    if 'compute' not in request.json['stages'][0]:
-        logging.error(f'Missing compute in stage')
-        return 'Missing compute in stages', 400
-    if 'input' not in request.json['stages'][0]:
-        logging.error(f'Missing input in stage')
-        return 'Missing input in stages', 400
-    if 'output' not in request.json['stages'][0]:
-        logging.error(f'Missing output in stage')
-        return 'Missing output in', 400
+
+    for _attr in ('algorithm', 'compute', 'input', 'output'):
+        if _attr not in stages[0]:
+            logging.error(f'Missing algorithm in stage')
+            return jsonify(error=f'Missing {_attr} in stage 0'), 400
+
     # check timeouts
-    timeout = os.getenv("ALGOPODTIMEOUT")
-    if timeout is not None:
-        timeout = int(timeout)
-        if timeout > 0:
-            if 'maxtime' in request.json['stages'][0]['compute']:
-                maxtime = int(request.json['stages'][0]['compute']['maxtime'])
-                if timeout < maxtime:
-                    request.json['stages'][0]['compute']['maxtime'] = timeout
-                    logging.error(f"Maxtime in workflow was {maxtime}. Overwritten to {timeout}")
-    agreement_id = request.json['agreementId']
-    owner = request.json['owner']
-    execution_id = generate_new_id()
-    logging.error(f'Got execution_id: {execution_id}')
-    body = create_execution(request.json, execution_id)
-    logging.error(f'Got body: {body}')
+    timeout = int(os.getenv("ALGOPODTIMEOUT", 0))
+    if timeout > 0:
+        if 'maxtime' in stages[0]['compute']:
+            maxtime = int(stages[0]['compute']['maxtime'])
+            if timeout < maxtime:
+                stages[0]['compute']['maxtime'] = timeout
+                logging.debug(f"Maxtime in workflow was {maxtime}. Overwritten to {timeout}")
+
+    job_id = generate_new_id()
+    logging.debug(f'Got job_id: {job_id}')
+    body = create_compute_job(
+        data, job_id, config.group, config.version, config.namespace
+    )
+    logging.debug(f'Got body: {body}')
+
+    kube_api = KubeAPI(config)
     try:
-        api_response = api_customobject.create_namespaced_custom_object(group, version, namespace,
-                                                                        plural, body)
+        api_response = kube_api.create_namespaced_custom_object(body)
         logging.info(api_response)
-        create_sql_job(agreement_id, str(execution_id), owner)
-        status_list = get_sql_status(agreement_id, str(execution_id), owner)
+        create_sql_job(agreement_id, str(job_id), owner)
+        status_list = get_sql_status(agreement_id, str(job_id), owner)
         return jsonify(status_list), 200
+
     except ApiException as e:
         logging.error(f'Exception when calling CustomObjectsApi->create_namespaced_custom_object: {e}')
         return 'Unable to create job', 400
 
 
 @services.route('/compute', methods=['PUT'])
-def stop_execution():
+def stop_compute_job():
     """
-    Stop the current workflow execution.
+    Stop the current compute job..
     ---
     tags:
       - operation
@@ -180,7 +172,7 @@ def stop_execution():
         type: string
       - name: jobId
         in: query
-        description: Id of the execution.
+        description: Id of the job.
         type: string
       - name: owner
         in: query
@@ -188,44 +180,44 @@ def stop_execution():
         type: string
     """
     try:
-        if 'agreementId' not in request.args:
+        data = request.args
+
+        agreement_id = data.get('agreementId')
+        owner = data.get('owner')
+        job_id = data.get('jobId')
+
+        if not agreement_id or len(agreement_id) < 2:
             agreement_id = None
-        elif len(request.args['agreementId']) < 2:
-            agreement_id = None
-        else:
-            agreement_id = request.args['agreementId']
-        if 'jobId' not in request.args:
+
+        if not job_id or len(job_id) < 2:
             job_id = None
-        elif len(request.args['jobId']) < 2:
-            job_id = None
-        else:
-            job_id = request.args['jobId']
-        if 'owner' not in request.args:
+
+        if not owner or len(owner) < 2:
             owner = None
-        elif len(request.args['owner']) < 2:
-            owner = None
-        else:
-            owner = request.args['owner']
+
         if owner is None and agreement_id is None and job_id is None:
-            logging.error(f'All args are null ?')
-            return f'You have to specify one of agreementId,jobId or owner', 400
+            msg = f'You have to specify one of agreementId, jobId or owner'
+            logging.error(msg)
+            return jsonify(error=msg), 400
         else:
-            jobslist = get_sql_jobs(agreement_id, job_id, owner)
-            for ajob in jobslist:
+            jobs_list = get_sql_jobs(agreement_id, job_id, owner)
+            for ajob in jobs_list:
                 name = ajob
-                logging.info(f'Stoping job : {name}')
+                logging.info(f'Stopping job : {name}')
                 stop_sql_job(name)
+
             status_list = get_sql_status(agreement_id, job_id, owner)
             return jsonify(status_list), 200
+
     except ApiException as e:
-        print("Exception when calling CustomObjectsApi->delete_namespaced_custom_object: %s\n" % e)
-        return 'Error stopping job', 400
+        logging.error(f'Exception when stopping compute job: {e}')
+        return f'Error stopping job: {e}', 400
 
 
 @services.route('/compute', methods=['DELETE'])
-def delete_execution():
+def delete_compute_job():
     """
-    Deletes the current workflow execution.
+    Deletes the current compute job.
     ---
     tags:
       - operation
@@ -238,7 +230,7 @@ def delete_execution():
         type: string
       - name: jobId
         in: query
-        description: Id of the execution.
+        description: Id of the job.
         type: string
       - name: owner
         in: query
@@ -259,48 +251,50 @@ def delete_execution():
     # default policy is decided by the existing finalizer set in the metadata.finalizers and the
     # resource-specific default policy. (optional)
     try:
-        if 'agreementId' not in request.args:
+        data = request.args
+        agreement_id = data.get('agreementId')
+        owner = data.get('owner')
+        job_id = data.get('jobId')
+
+        if not agreement_id or len(agreement_id) < 2:
             agreement_id = None
-        elif len(request.args['agreementId']) < 2:
-            agreement_id = None
-        else:
-            agreement_id = request.args['agreementId']
-        if 'jobId' not in request.args:
+
+        if not job_id or len(job_id) < 2:
             job_id = None
-        elif len(request.args['jobId']) < 2:
-            job_id = None
-        else:
-            job_id = request.args['jobId']
-        if 'owner' not in request.args:
+
+        if not owner or len(owner) < 2:
             owner = None
-        elif len(request.args['owner']) < 2:
-            owner = None
-        else:
-            owner = request.args['owner']
+
         if owner is None and agreement_id is None and job_id is None:
-            logging.error(f'All args are null ?')
-            return f'You have to specify one of agreementId,jobId or owner', 400
+            msg = f'You have to specify one of agreementId, jobId or owner'
+            logging.error(msg)
+            return jsonify(error=msg), 400
         else:
-            jobslist = get_sql_jobs(agreement_id, job_id, owner)
-            for ajob in jobslist:
+            kube_api = KubeAPI(config)
+            jobs_list = get_sql_jobs(agreement_id, job_id, owner)
+            for ajob in jobs_list:
                 name = ajob
                 logging.info(f'Deleting job : {name}')
                 remove_sql_job(name)
-                api_response = api_customobject.delete_namespaced_custom_object(group, version, namespace,
-                                                                                plural, name, body,
-                                                                                grace_period_seconds=grace_period_seconds,
-                                                                                orphan_dependents=orphan_dependents,
-                                                                                propagation_policy=propagation_policy)
+                api_response = kube_api.delete_namespaced_custom_object(
+                    name,
+                    body,
+                    grace_period_seconds=grace_period_seconds,
+                    orphan_dependents=orphan_dependents,
+                    propagation_policy=propagation_policy
+                )
                 logging.info(api_response)
+
         status_list = get_sql_status(agreement_id, job_id, owner)
         return jsonify(status_list), 200
+
     except ApiException as e:
         logging.error(f'Exception when calling CustomObjectsApi->delete_namespaced_custom_object: {e}')
-        return 'Error deleteing job', 400
+        return jsonify(error=f'Error deleting job {e}'), 400
 
 
 @services.route('/compute', methods=['GET'])
-def get_execution_status():
+def get_compute_job_status():
     """
     Get status for an specific or multiple jobs.
     ---
@@ -315,7 +309,7 @@ def get_execution_status():
         type: string
       - name: jobId
         in: query
-        description: Id of the execution.
+        description: Id of the job.
         type: string
       - name: owner
         in: query
@@ -328,379 +322,30 @@ def get_execution_status():
         description: Error
     """
     try:
-        if 'agreementId' not in request.args:
+        data = request.args
+        agreement_id = data.get('agreementId')
+        owner = data.get('owner')
+        job_id = data.get('jobId')
+
+        if not agreement_id or len(agreement_id) < 2:
             agreement_id = None
-        elif len(request.args['agreementId']) < 2:
-            agreement_id = None
-        else:
-            agreement_id = request.args['agreementId']
-        if 'jobId' not in request.args:
+
+        if not job_id or len(job_id) < 2:
             job_id = None
-        elif len(request.args['jobId']) < 2:
-            job_id = None
-        else:
-            job_id = request.args['jobId']
-        if 'owner' not in request.args:
+
+        if not owner or len(owner) < 2:
             owner = None
-        elif len(request.args['owner']) < 2:
-            owner = None
-        else:
-            owner = request.args['owner']
+
         if owner is None and agreement_id is None and job_id is None:
-            logging.error(f'All args are null ?')
-            return f'You have to specify one of agreementId,jobId or owner', 400
+            msg = f'You have to specify one of agreementId, jobId or owner'
+            logging.error(msg)
+            return jsonify(error=msg), 400
         else:
-            logging.error("Try to start")
+            logging.debug("Try to start")
             api_response = get_sql_status(agreement_id, job_id, owner)
-            # logging.error(f'Got from sql: {api_response}')
             return jsonify(api_response), 200
-    except ApiException as e:
-        logging.error(
-            f'Exception when calling status: {e}')
-        return 'Error getting the status', 400
-
-
-# methods not in API.md - admin only
-
-@services.route('/pgsqlinit', methods=['POST'])
-def init_pgsqlexecution():
-    """
-    Init pgsql database
-    ---
-    tags:
-      - operation
-    consumes:
-      - application/json
-    """
-    output = ""
-    connection = None
-    cursor = None
-    try:
-        connection = psycopg2.connect(user=os.getenv("POSTGRES_USER"),
-                                      password=os.getenv("POSTGRES_PASSWORD"),
-                                      host=os.getenv("POSTGRES_HOST"),
-                                      port=os.getenv("POSTGRES_PORT"),
-                                      database=os.getenv("POSTGRES_DB"))
-        cursor = connection.cursor()
-        create_table_query = '''CREATE TABLE IF NOT EXISTS jobs 
-          (agreementId           varchar(255) NOT NULL,
-          workflowId         varchar(255) NOT NULL,
-          owner         varchar(255),
-          status  int,
-          statusText varchar(255),
-          dateCreated timestamp without time zone default NOW(),
-          dateFinished timestamp without time zone default NULL,
-          configlogURL text,
-          publishlogURL text,
-          algologURL text,
-          outputsURL text,
-          ddo text,
-          namespace varchar(255),
-          stopreq smallint default 0,
-          removed smallint default 0
-          ); '''
-        cursor.execute(create_table_query)
-        # queries below are for upgrade purposes
-        create_table_query = '''ALTER TABLE jobs ADD COLUMN IF NOT EXISTS namespace varchar(255)'''
-        cursor.execute(create_table_query)
-        create_table_query = '''ALTER TABLE jobs ADD COLUMN IF NOT EXISTS stopreq smallint default 0'''
-        cursor.execute(create_table_query)
-        create_table_query = '''ALTER TABLE jobs ADD COLUMN IF NOT EXISTS removed smallint default 0'''
-        cursor.execute(create_table_query)
-        create_index_query = '''CREATE unique INDEX IF NOT EXISTS uniq_agreementId_workflowId ON jobs (agreementId,workflowId)'''
-        cursor.execute(create_index_query)
-        connection.commit()
-    except (Exception, psycopg2.Error) as error:
-        output = output + "Error PostgreSQL:" + str(error)
-    finally:
-        # closing database connection.
-        if connection and cursor:
-            cursor.close()
-            connection.close()
-
-    return output, 200
-
-
-@services.route('/info', methods=['GET'])
-def get_execution_info():
-    """
-    Get info for an execution id.
-    ---
-    tags:
-      - operation
-    consumes:
-      - application/json
-    parameters:
-      - name: jobId
-        in: query
-        description: Id of the execution.
-        required: true
-        type: string
-    """
-    try:
-        execution_id = request.args['jobId']
-        api_response = api_customobject.get_namespaced_custom_object(group, version, namespace, plural,
-                                                                     execution_id)
-        logging.info(api_response)
-        return jsonify(api_response), 200
-    except ApiException as e:
-        logging.error(f'The executionId {execution_id} is not registered in your namespace.')
-        return f'The executionId {execution_id} is not registered in your namespace.', 400
-
-
-@services.route('/list', methods=['GET'])
-def list_executions():
-    """
-    List all the execution workflows.
-    ---
-    tags:
-      - operation
-    consumes:
-      - application/json
-    """
-    try:
-        api_response = api_customobject.list_namespaced_custom_object(group, version, namespace, plural)
-        result = list()
-        for i in api_response['items']:
-            result.append(i['metadata']['name'])
-        logging.info(api_response)
-        return jsonify(result), 200
 
     except ApiException as e:
-        logging.error(
-            f'Exception when calling CustomObjectsApi->list_cluster_custom_object: {e}')
-        return 'Error listing workflows', 400
-
-
-@services.route('/logs', methods=['GET'])
-def get_logs():
-    """
-    Get the logs for an execution id.
-    ---
-    tags:
-      - operation
-    consumes:
-      - text/plain
-    parameters:
-      - name: jobId
-        in: query
-        description: Id of the execution.
-        required: true
-        type: string
-      - name: component
-        in: query
-        description: Workflow component (configure, algorithm, publish)
-        required: true
-        type: string
-    responses:
-      200:
-        description: Get correctly the logs
-      400:
-        description: Error consume Kubernetes API
-      404:
-        description: Pod not found for the given parameters
-    """
-    data = request.args
-    try:
-        execution_id = data.get('jobId')
-        component = data.get('component')
-        # First we need to get the name of the pods
-        label_selector = f'workflow={execution_id},component={component}'
-        logging.debug(f'Looking pods in ns {namespace} with labels {label_selector}')
-        pod_response = api_core.list_namespaced_pod(namespace, label_selector=label_selector)
-    except ApiException as e:
-        logging.error(
-            f'Exception when calling CustomObjectsApi->list_namespaced_pod: {e}')
-        return 'Error getting the logs', 400
-
-    try:
-        pod_name = pod_response.items[0].metadata.name
-        logging.debug(f'pods found: {pod_response}')
-    except IndexError as e:
-        logging.warning(f'Exception getting information about the pod with labels {label_selector}.'
-                        f' Probably pod does not exist')
-        return f'Pod with workflow={execution_id} and component={component} not found', 404
-
-    try:
-        logging.debug(f'looking logs for pod {pod_name} in namespace {namespace}')
-        logs_response = api_core.read_namespaced_pod_log(name=pod_name, namespace=namespace)
-        r = Response(response=logs_response, status=200, mimetype="text/plain")
-        r.headers["Content-Type"] = "text/plain; charset=utf-8"
-        return r
-
-    except ApiException as e:
-        logging.error(
-            f'Exception when calling CustomObjectsApi->read_namespaced_pod_log: {e}')
-        return 'Error getting the logs', 400
-
-
-def create_execution(workflow, execution_id):
-    execution = dict()
-    execution['apiVersion'] = group + '/' + version
-    execution['kind'] = 'WorkFlow'
-    execution['metadata'] = dict()
-    execution['metadata']['name'] = execution_id
-    execution['metadata']['namespace'] = namespace
-    execution['metadata']['labels'] = dict()
-    execution['metadata']['labels']['workflow'] = execution_id
-    execution['spec'] = dict()
-    execution['spec']['metadata'] = workflow
-    return execution
-
-
-# TODO Use the commons utils library to do this when we set up the project.
-def generate_new_id():
-    """
-    Generate a new id without prefix.
-
-    :return: Id, str
-    """
-    return uuid.uuid4().hex
-
-
-# sql handlers
-
-def get_sql_status(agreement_id, job_id, owner):
-    # enforce strings
-    logging.debug("Start get_sql_status\n")
-    logging.debug("Connected\n")
-    params = dict()
-    select_query = '''
-    SELECT agreementId, workflowId, owner, status, statusText, 
-        extract(epoch from dateCreated) as dateCreated, 
-        extract(epoch from dateFinished) as dateFinished,
-        configlogURL, publishlogURL, algologURL, outputsURL, ddo, stopreq, removed 
-    FROM jobs WHERE 1=1
-    '''
-
-    if agreement_id is not None:
-        select_query = select_query + " AND agreementId=%(agreementId)s"
-        params['agreementId'] = str(agreement_id)
-
-    if job_id is not None:
-        select_query = select_query + " AND workflowId=%(jobId)s"
-        params['jobId'] = str(job_id)
-
-    if owner is not None:
-        select_query = select_query + " AND owner=%(owner)s"
-        params['owner'] = str(owner)
-
-    logging.debug(f'Got select_query: {select_query}')
-    logging.debug(f'Got params: {params}')
-
-    result = []
-    rows = _execute_query(select_query, params, 'get_sql_status', get_rows=True)
-    if not rows:
-        return
-
-    for row in rows:
-        temprow = dict()
-        temprow['agreementId'] = row[0]
-        temprow['jobId'] = row[1]
-        temprow['owner'] = row[2]
-        temprow['status'] = row[3]
-        temprow['statusText'] = row[4]
-        temprow['dateCreated'] = row[5]
-        temprow['dateFinished'] = row[6]
-        # temprow['configlogUrl']=row[7]
-        # temprow['publishlogUrl']=row[8]
-        temprow['algologUrl'] = row[9]
-        temprow['outputsUrl'] = row[10]
-        temprow['ddo'] = ''
-        temprow['did'] = ''
-        if row[11] is not None:
-            temprow['ddo'] = str(row[11])
-            ddo_json = json.loads(temprow['ddo'])
-            if 'id' in ddo_json:
-                temprow['did'] = ddo_json['id']
-        temprow['stopreq'] = row[12]
-        temprow['removed'] = row[13]
-        result.append(temprow)
-    return result
-
-
-def get_sql_jobs(agreement_id, job_id, owner):
-    logging.debug("Start get_sql_status\n")
-    logging.debug("Connected\n")
-    params = dict()
-    select_query = "SELECT workflowId FROM jobs WHERE 1=1"
-    if agreement_id is not None:
-        select_query = select_query + " AND agreementId=%(agreementId)s"
-        params['agreementId'] = str(agreement_id)
-    if job_id is not None:
-        select_query = select_query + " AND workflowId=%(jobId)s"
-        params['jobId'] = str(job_id)
-    if owner is not None:
-        select_query = select_query + " AND owner=%(owner)s"
-        params['owner'] = str(owner)
-    logging.error(f'Got select_query: {select_query}')
-    logging.error(f'Got params: {params}')
-    rows = _execute_query(select_query, params, 'get_sql_jobs', get_rows=True)
-    if not rows:
-        return
-
-    return [row[0] for row in rows]
-
-
-def create_sql_job(agreement_id, job_id, owner):
-    postgres_insert_query = """
-        INSERT 
-            INTO jobs 
-                (agreementId,workflowId,owner,status,statusText) 
-            VALUES 
-                (%s, %s, %s, %s, %s)"""
-    record_to_insert = (str(agreement_id), str(job_id), str(owner), 10, "Job started")
-    return _execute_query(postgres_insert_query, record_to_insert, 'create_sql_job')
-
-
-def stop_sql_job(execution_id):
-    postgres_insert_query = """UPDATE jobs SET stopreq=1 WHERE workflowId=%s"""
-    record_to_insert = (str(execution_id),)
-    return _execute_query(postgres_insert_query, record_to_insert, 'stop_sql_job')
-
-
-def remove_sql_job(execution_id):
-    postgres_insert_query = """ UPDATE jobs SET removed=1 WHERE workflowId=%s"""
-    record_to_insert = (str(execution_id),)
-    return _execute_query(postgres_insert_query, record_to_insert, 'remove_sql_job')
-
-
-def get_pg_connection_and_cursor():
-    try:
-        connection = psycopg2.connect(user=os.getenv("POSTGRES_USER"),
-                                      password=os.getenv("POSTGRES_PASSWORD"),
-                                      host=os.getenv("POSTGRES_HOST"),
-                                      port=os.getenv("POSTGRES_PORT"),
-                                      database=os.getenv("POSTGRES_DB"))
-        connection.set_client_encoding('LATIN9')
-        cursor = connection.cursor()
-        return connection, cursor
-    except (Exception, psycopg2.Error) as error:
-        logging.error(f'New PG connect error: {error}')
-        return None, None
-
-
-def _execute_query(query, record, msg, get_rows=False):
-    connection, cursor = get_pg_connection_and_cursor()
-    if not connection or not cursor:
-        return
-
-    try:
-        cursor.execute(query, record)
-        if get_rows:
-            result = []
-            row = cursor.fetchone()
-            while row is not None:
-                result.append(row[0])
-                row = cursor.fetchone()
-            return result
-        else:
-            connection.commit()
-
-    except (Exception, psycopg2.Error) as error:
-        logging.error(f'Got PG error in {msg}: {error}')
-    finally:
-        # closing database connection.
-        if connection:
-            cursor.close()
-            connection.close()
+        msg = f'Error getting the status: {e}'
+        logging.error(msg)
+        return msg, 400
