+++ operator_service/utils.py
@@ -1,0 +1,32 @@
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+def create_compute_job(workflow, execution_id, group, version, namespace):
+    execution = dict()
+    execution['apiVersion'] = group + '/' + version
+    execution['kind'] = 'WorkFlow'
+    execution['metadata'] = dict()
+    execution['metadata']['name'] = execution_id
+    execution['metadata']['namespace'] = namespace
+    execution['metadata']['labels'] = dict()
+    execution['metadata']['labels']['workflow'] = execution_id
+    execution['spec'] = dict()
+    execution['spec']['metadata'] = workflow
+    return execution
+
+
+def check_required_attributes(required_attributes, data, method):
+    assert isinstance(data, dict), 'invalid payload format.'
+    logger.info('got %s request: %s' % (method, data))
+    if not data:
+        logger.error('%s request failed: data is empty.' % method)
+        return 'payload seems empty.', 400
+
+    for attr in required_attributes:
+        if attr not in data:
+            logger.error('%s request failed: required attr %s missing.' % (method, attr))
+            return '"%s" is required in the call to %s' % (attr, method), 400
+
+    return None, None
