+++ test/test_admin.py
@@ -1,0 +1,15 @@
+
+def test_init_pgsqlexecution():
+    return
+
+
+def test_get_execution_info():
+    return
+
+
+def test_list_executions():
+    return
+
+
+def test_get_logs():
+    return
